requirements.txt
requests==2.31.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.17.0
python-dateutil==2.8.2